// src/lib/gramjs/tl/api.d.ts
namespace Api {
// ...
  type bytes = Buffer;
// ...
  export type TypeUser = UserEmpty | User;
// ...
  export type TypeUpdate = UpdateNewMessage | UpdateMessageID | UpdateDeleteMessages | UpdateUserTyping | UpdateUser | UpdateChatUserTyping | UpdateChatParticipants | UpdateUserStatus | UpdateUserName | UpdateNewAuthorization | UpdateUserPhoto | UpdateNewEncryptedMessage | UpdateEncryptedChatTyping | UpdateEncryption | UpdateEncryptedMessagesRead | UpdateChatParticipantAdd | UpdateChatParticipantDelete | UpdateDcOptions | UpdateNotifySettings | UpdateServiceNotification | UpdatePrivacy | UpdateUserPhone | UpdateReadHistoryInbox | UpdateReadHistoryOutbox | UpdateWebPage | UpdateReadMessagesContents | UpdateChannelTooLong | UpdateChannel | UpdateNewChannelMessage | UpdateReadChannelInbox | UpdateDeleteChannelMessages | UpdateChannelMessageViews | UpdateChatParticipantAdmin | UpdateNewStickerSet | UpdateStickerSetsOrder | UpdateStickerSets | UpdateSavedGifs | UpdateBotInlineQuery | UpdateBotInlineSend | UpdateEditChannelMessage | UpdateBotCallbackQuery | UpdateEditMessage | UpdateInlineBotCallbackQuery | UpdateReadChannelOutbox | UpdateDraftMessage | UpdateReadFeaturedStickers | UpdateRecentStickers | UpdateConfig | UpdatePtsChanged | UpdateChannelWebPage | UpdateDialogPinned | UpdatePinnedDialogs | UpdateBotWebhookJSON | UpdateBotWebhookJSONQuery | UpdateBotShippingQuery | UpdateBotPrecheckoutQuery | UpdatePhoneCall | UpdateLangPackTooLong | UpdateLangPack | UpdateFavedStickers | UpdateChannelReadMessagesContents | UpdateContactsReset | UpdateChannelAvailableMessages | UpdateDialogUnreadMark | UpdateMessagePoll | UpdateChatDefaultBannedRights | UpdateFolderPeers | UpdatePeerSettings | UpdatePeerLocated | UpdateNewScheduledMessage | UpdateDeleteScheduledMessages | UpdateTheme | UpdateGeoLiveViewed | UpdateLoginToken | UpdateMessagePollVote | UpdateDialogFilter | UpdateDialogFilterOrder | UpdateDialogFilters | UpdatePhoneCallSignalingData | UpdateChannelMessageForwards | UpdateReadChannelDiscussionInbox | UpdateReadChannelDiscussionOutbox | UpdatePeerBlocked | UpdateChannelUserTyping | UpdatePinnedMessages | UpdatePinnedChannelMessages | UpdateChat | UpdateGroupCallParticipants | UpdateGroupCall | UpdatePeerHistoryTTL | UpdateChatParticipant | UpdateChannelParticipant | UpdateBotStopped | UpdateGroupCallConnection | UpdateBotCommands | UpdatePendingJoinRequests | UpdateBotChatInviteRequester | UpdateMessageReactions | UpdateAttachMenuBots | UpdateWebViewResultSent | UpdateBotMenuButton | UpdateSavedRingtones | UpdateTranscribedAudio | UpdateReadFeaturedEmojiStickers | UpdateUserEmojiStatus | UpdateRecentEmojiStatuses | UpdateRecentReactions | UpdateMoveStickerSetToTop | UpdateMessageExtendedMedia;
// ...
  export class UserEmpty extends VirtualClass<{
    id: long;
  }> {
    id: long;
  };
  export class User extends VirtualClass<{
    // flags: undefined;
    business?: true;
    public?: true;
    self?: true;
    contact?: true;
    mutualContact?: true;
    deleted?: true;
    bot?: true;
    botChatHistory?: true;
    botNochats?: true;
    verified?: true;
    restricted?: true;
    min?: true;
    botInlineGeo?: true;
    support?: true;
    scam?: true;
    applyMinPhoto?: true;
    fake?: true;
    botAttachMenu?: true;
    premium?: true;
    attachMenuEnabled?: true;
    id: long;
    accessHash?: long;
    firstName?: string;
    lastName?: string;
    username?: string;
    phone?: string;
    photo?: Api.TypeUserProfilePhoto;
    status?: Api.TypeUserStatus;
    botInfoVersion?: int;
    restrictionReason?: Api.TypeRestrictionReason[];
    botInlinePlaceholder?: string;
    langCode?: string;
    emojiStatus?: Api.TypeEmojiStatus;
    botDescription?: string;
  }> {
    // flags: undefined;
    business?: true;
    public?: true;
    self?: true;
    contact?: true;
    mutualContact?: true;
    deleted?: true;
    bot?: true;
    botChatHistory?: true;
    botNochats?: true;
    verified?: true;
    restricted?: true;
    min?: true;
    botInlineGeo?: true;
    support?: true;
    scam?: true;
    applyMinPhoto?: true;
    fake?: true;
    botAttachMenu?: true;
    premium?: true;
    attachMenuEnabled?: true;
    id: long;
    accessHash?: long;
    firstName?: string;
    lastName?: string;
    username?: string;
    phone?: string;
    photo?: Api.TypeUserProfilePhoto;
    status?: Api.TypeUserStatus;
    botInfoVersion?: int;
    restrictionReason?: Api.TypeRestrictionReason[];
    botInlinePlaceholder?: string;
    langCode?: string;
    emojiStatus?: Api.TypeEmojiStatus;
    botDescription?: string;
  };
// ...
  export class UpdateNewMessage extends VirtualClass<{
    message: Api.TypeMessage;
    pts: int;
    ptsCount: int;
  }> {
    message: Api.TypeMessage;
    pts: int;
    ptsCount: int;
  };
  export class UpdateMessageID extends VirtualClass<{
    id: int;
    randomId: long;
  }> {
    id: int;
    randomId: long;
  };
  export class UpdateDeleteMessages extends VirtualClass<{
    messages: int[];
    pts: int;
    ptsCount: int;
  }> {
    messages: int[];
    pts: int;
    ptsCount: int;
  };
  export class UpdateUserTyping extends VirtualClass<{
    userId: long;
    action: Api.TypeSendMessageAction;
  }> {
    userId: long;
    action: Api.TypeSendMessageAction;
  };
  export class UpdateUser extends VirtualClass<{
    userId: long;
  }> {
    userId: long;
  };
  export class UpdateChatUserTyping extends VirtualClass<{
    chatId: long;
    fromId: Api.TypePeer;
    action: Api.TypeSendMessageAction;
  }> {
    chatId: long;
    fromId: Api.TypePeer;
    action: Api.TypeSendMessageAction;
  };
  export class UpdateChatParticipants extends VirtualClass<{
    participants: Api.TypeChatParticipants;
  }> {
    participants: Api.TypeChatParticipants;
  };
  export class UpdateUserStatus extends VirtualClass<{
    userId: long;
    status: Api.TypeUserStatus;
  }> {
    userId: long;
    status: Api.TypeUserStatus;
  };
  export class UpdateUserName extends VirtualClass<{
    userId: long;
    firstName: string;
    lastName: string;
    username: string;
  }> {
    userId: long;
    firstName: string;
    lastName: string;
    username: string;
  };
  export class UpdateNewAuthorization extends VirtualClass<{
    // flags: undefined;
    unconfirmed?: true;
    hash: long;
    date?: int;
    device?: string;
    location?: string;
  }> {
    // flags: undefined;
    unconfirmed?: true;
    hash: long;
    date?: int;
    device?: string;
    location?: string;
  };
  export class UpdateUserPhoto extends VirtualClass<{
    userId: long;
    date: int;
    photo: Api.TypeUserProfilePhoto;
    previous: Bool;
  }> {
    userId: long;
    date: int;
    photo: Api.TypeUserProfilePhoto;
    previous: Bool;
  };
  export class UpdateNewEncryptedMessage extends VirtualClass<{
    message: Api.TypeEncryptedMessage;
    qts: int;
  }> {
    message: Api.TypeEncryptedMessage;
    qts: int;
  };
  export class UpdateEncryptedChatTyping extends VirtualClass<{
    chatId: int;
  }> {
    chatId: int;
  };
  export class UpdateEncryption extends VirtualClass<{
    chat: Api.TypeEncryptedChat;
    date: int;
  }> {
    chat: Api.TypeEncryptedChat;
    date: int;
  };
  export class UpdateEncryptedMessagesRead extends VirtualClass<{
    chatId: int;
    maxDate: int;
    date: int;
  }> {
    chatId: int;
    maxDate: int;
    date: int;
  };
  export class UpdateChatParticipantAdd extends VirtualClass<{
    chatId: long;
    userId: long;
    inviterId: long;
    date: int;
    version: int;
  }> {
    chatId: long;
    userId: long;
    inviterId: long;
    date: int;
    version: int;
  };
  export class UpdateChatParticipantDelete extends VirtualClass<{
    chatId: long;
    userId: long;
    version: int;
  }> {
    chatId: long;
    userId: long;
    version: int;
  };
  export class UpdateDcOptions extends VirtualClass<{
    dcOptions: Api.TypeDcOption[];
  }> {
    dcOptions: Api.TypeDcOption[];
  };
  export class UpdateNotifySettings extends VirtualClass<{
    peer: Api.TypeNotifyPeer;
    notifySettings: Api.TypePeerNotifySettings;
  }> {
    peer: Api.TypeNotifyPeer;
    notifySettings: Api.TypePeerNotifySettings;
  };
  export class UpdateServiceNotification extends VirtualClass<{
    // flags: undefined;
    popup?: true;
    inboxDate?: int;
    type: string;
    message: string;
    media: Api.TypeMessageMedia;
    entities: Api.TypeMessageEntity[];
  }> {
    // flags: undefined;
    popup?: true;
    inboxDate?: int;
    type: string;
    message: string;
    media: Api.TypeMessageMedia;
    entities: Api.TypeMessageEntity[];
  };
  export class UpdatePrivacy extends VirtualClass<{
    key: Api.TypePrivacyKey;
    rules: Api.TypePrivacyRule[];
  }> {
    key: Api.TypePrivacyKey;
    rules: Api.TypePrivacyRule[];
  };
  export class UpdateUserPhone extends VirtualClass<{
    userId: long;
    phone: string;
  }> {
    userId: long;
    phone: string;
  };
  export class UpdateReadHistoryInbox extends VirtualClass<{
    // flags: undefined;
    folderId?: int;
    peer: Api.TypePeer;
    maxId: int;
    stillUnreadCount: int;
    pts: int;
    ptsCount: int;
  }> {
    // flags: undefined;
    folderId?: int;
    peer: Api.TypePeer;
    maxId: int;
    stillUnreadCount: int;
    pts: int;
    ptsCount: int;
  };
  export class UpdateReadHistoryOutbox extends VirtualClass<{
    peer: Api.TypePeer;
    maxId: int;
    pts: int;
    ptsCount: int;
  }> {
    peer: Api.TypePeer;
    maxId: int;
    pts: int;
    ptsCount: int;
  };
  export class UpdateWebPage extends VirtualClass<{
    webpage: Api.TypeWebPage;
    pts: int;
    ptsCount: int;
  }> {
    webpage: Api.TypeWebPage;
    pts: int;
    ptsCount: int;
  };
  export class UpdateReadMessagesContents extends VirtualClass<{
    messages: int[];
    pts: int;
    ptsCount: int;
  }> {
    messages: int[];
    pts: int;
    ptsCount: int;
  };
  export class UpdateChannelTooLong extends VirtualClass<{
    // flags: undefined;
    channelId: long;
    pts?: int;
  }> {
    // flags: undefined;
    channelId: long;
    pts?: int;
  };
  export class UpdateChannel extends VirtualClass<{
    channelId: long;
  }> {
    channelId: long;
  };
  export class UpdateNewChannelMessage extends VirtualClass<{
    message: Api.TypeMessage;
    pts: int;
    ptsCount: int;
  }> {
    message: Api.TypeMessage;
    pts: int;
    ptsCount: int;
  };
  export class UpdateReadChannelInbox extends VirtualClass<{
    // flags: undefined;
    folderId?: int;
    channelId: long;
    maxId: int;
    stillUnreadCount: int;
    pts: int;
  }> {
    // flags: undefined;
    folderId?: int;
    channelId: long;
    maxId: int;
    stillUnreadCount: int;
    pts: int;
  };
  export class UpdateDeleteChannelMessages extends VirtualClass<{
    channelId: long;
    messages: int[];
    pts: int;
    ptsCount: int;
  }> {
    channelId: long;
    messages: int[];
    pts: int;
    ptsCount: int;
  };
  export class UpdateChannelMessageViews extends VirtualClass<{
    channelId: long;
    id: int;
    views: int;
  }> {
    channelId: long;
    id: int;
    views: int;
  };
  export class UpdateChatParticipantAdmin extends VirtualClass<{
    chatId: long;
    userId: long;
    isAdmin: Bool;
    version: int;
  }> {
    chatId: long;
    userId: long;
    isAdmin: Bool;
    version: int;
  };
// ...
  export class UpdateLoginToken extends VirtualClass<void> {};
// ...
  export class CodeSettings extends VirtualClass<{
    // flags: undefined;
    allowFlashcall?: true;
    currentNumber?: true;
    allowAppHash?: true;
    allowMissedCall?: true;
    logoutTokens?: bytes[];
  } | void> {
    // flags: undefined;
    allowFlashcall?: true;
    currentNumber?: true;
    allowAppHash?: true;
    allowMissedCall?: true;
    logoutTokens?: bytes[];
  };
// ...
  export namespace account {
// ...
    export class UpdateNotifySettings extends Request<Partial<{
      peer: Api.TypeInputNotifyPeer;
      settings: Api.TypeInputPeerNotifySettings;
    }>, Bool> {
      peer: Api.TypeInputNotifyPeer;
      settings: Api.TypeInputPeerNotifySettings;
    };
// ...
  }
// ...
}